Split master integration job tests into three groups
