+++ .travis.yml
@@ -31,8 +31,12 @@ matrix:
       env: TITUS_TEST_GROUP=unit
     - name: "Integration Tests (exclude master)"
       env: TITUS_TEST_GROUP=integrationExcludeMaster
-    - name: "Integration Tests (only master)"
-      env: TITUS_TEST_GROUP=integrationOnlyMaster
+    - name: "Integration Tests (only master 1)"
+      env: TITUS_TEST_GROUP=integrationOnlyMaster1
+    - name: "Integration Tests (only master 2)"
+      env: TITUS_TEST_GROUP=integrationOnlyMaster2
+    - name: "Integration Tests (only master 3)"
+      env: TITUS_TEST_GROUP=integrationOnlyMaster3
     - name: "Integration Tests (not parallelizable)"
       env: TITUS_TEST_GROUP=integrationNotParallelizable
     - name: "Docker Compose Integration"
