+++ build.gradle
@@ -190,8 +190,14 @@ subprojects {
 
         if( "excludeMaster".equalsIgnoreCase(integrationTestScope)) {
             exclude 'com/netflix/titus/master/integration/**/*'
-        } else if( "onlyMaster".equalsIgnoreCase(integrationTestScope)) {
+        } else if( "onlyMaster1".equalsIgnoreCase(integrationTestScope)) {
             include 'com/netflix/titus/master/integration/**/*'
+            exclude 'com/netflix/titus/master/integration/v3/job/**/*'
+        } else if( "onlyMaster2".equalsIgnoreCase(integrationTestScope)) {
+            include 'com/netflix/titus/master/integration/v3/job/**/*'
+            exclude 'com/netflix/titus/master/integration/v3/job/query/*'
+        } else if( "onlyMaster3".equalsIgnoreCase(integrationTestScope)) {
+            include 'com/netflix/titus/master/integration/v3/job/query/*'
         }
 
         useJUnit {
